Move out tests for Workflow API calls
